Update fasttrack-dynamics-365-overview.md
I have added a few clarifications, primarily in the area of Solution Blueprint Review.   The topics are now aligned with the agenda.   The time frame was updated to 8 hours from 2-4.    And commentary was added to indicate that this is the first step in the engagement and other subsequent steps will not start until it is complete.